Finished 3x3 TTT RL Alg.
A working RL algorithm which can train an agent to play tic-tac-toe of board size up to 3x3. The agent plays against itself during the training phase, making moves based on an epsilon-greedy approach. The Q-table is implemented as a 2d array, which is the reason the board can't be larger than 3x3, as the array gets intractably large otherwise. The reward scheme is as follows: receive a reward if the agent plays a winning move, and receive a negative reward if the agent plays a move such that the opponent can win next turn.  After training, the agent plays against itself, choosing moves based on the maximum Q-value in the Q-table. These games are printed out to the screen for viewing.

The agent will play winning moves if available, and is smart enough to also stop all immediate threats his opponent poses. However, after 250,000 moves the agent didn't learn to stop all double attack moves. May need to adjust hyperparameters or let the algorithm train longer.